+++ objc2/src/declare.rs
@@ -10,9 +10,9 @@ The following example demonstrates declaring a class named `MyNumber` that has
 one ivar, a `u32` named `_number` and a `number` method that returns it:
 
 ``` no_run
-# use objc::{class, sel};
-# use objc::declare::ClassDecl;
-# use objc::runtime::{Class, Object, Sel};
+# use objc2::{class, sel};
+# use objc2::declare::ClassDecl;
+# use objc2::runtime::{Class, Object, Sel};
 let superclass = class!(NSObject);
 let mut decl = ClassDecl::new("MyNumber", superclass).unwrap();