+++ objc2/tests-ios/prelude.rs
@@ -1,9 +1,9 @@
 #[macro_use]
-extern crate objc;
+extern crate objc2;
 
-use objc::rc::*;
-use objc::runtime::*;
-pub use objc::*;
+use objc2::rc::*;
+use objc2::runtime::*;
+pub use objc2::*;
 
 #[path = "../src/test_utils.rs"]
 mod test_utils;