+++ objc2_foundation/src/enumerator.rs
@@ -4,8 +4,8 @@ use core::ptr;
 use core::slice;
 use std::os::raw::c_ulong;
 
-use objc::runtime::Object;
-use objc::{msg_send, Encode, Encoding, RefEncode};
+use objc2::runtime::Object;
+use objc2::{msg_send, Encode, Encoding, RefEncode};
 use objc2_id::Id;
 
 use super::INSObject;